move decode ops to tf package
