+++ spark/dl/src/main/scala/com/intel/analytics/bigdl/utils/serializer/ModuleSerializer.scala
@@ -19,8 +19,8 @@ import com.intel.analytics.bigdl.Module
 import com.intel.analytics.bigdl.nn._
 import com.intel.analytics.bigdl.nn.abstractnn.{AbstractModule, Activity, TensorModule}
 import com.intel.analytics.bigdl.nn.keras.{KerasLayer, KerasLayerSerializer, Model, Sequential => KSequential}
-import com.intel.analytics.bigdl.nn.ops.{DecodeRawSerializer, RandomUniform => RandomUniformOps}
-import com.intel.analytics.bigdl.nn.tf.{StrideSlice, ParseExample}
+import com.intel.analytics.bigdl.nn.ops.{RandomUniform => RandomUniformOps}
+import com.intel.analytics.bigdl.nn.tf.{StrideSlice, ParseExample, DecodeRawSerializer}
 import com.intel.analytics.bigdl.optim.Regularizer
 import com.intel.analytics.bigdl.tensor.Tensor
 import com.intel.analytics.bigdl.tensor.TensorNumericMath.TensorNumeric
@@ -209,7 +209,7 @@ object ModuleSerializer extends ModuleSerializable{
       quantized.Linear)
     registerModule("com.intel.analytics.bigdl.nn.tf.ParseExample", ParseExample)
     registerModule("com.intel.analytics.bigdl.nn.SReLU", SReLU)
-    registerModule("com.intel.analytics.bigdl.nn.ops.DecodeRaw", DecodeRawSerializer)
+    registerModule("com.intel.analytics.bigdl.nn.tf.DecodeRaw", DecodeRawSerializer)
     registerModule("com.intel.analytics.bigdl.nn.ops.RandomUniform", RandomUniformOps)
     registerModule("com.intel.analytics.bigdl.nn.tf.StrideSlice", StrideSlice)
     registerModule("com.intel.analytics.bigdl.nn.MultiRNNCell", MultiRNNCell)
