+++ spark/dl/src/main/scala/com/intel/analytics/bigdl/utils/tf/loaders/DecodePng.scala
@@ -18,7 +18,7 @@ package com.intel.analytics.bigdl.utils.tf.loaders
 import java.nio.ByteOrder
 
 import com.intel.analytics.bigdl.Module
-import com.intel.analytics.bigdl.nn.ops.{DecodePng => DecodePngOp}
+import com.intel.analytics.bigdl.nn.tf.{DecodePng => DecodePngOp}
 import com.intel.analytics.bigdl.tensor.TensorNumericMath.TensorNumeric
 import com.intel.analytics.bigdl.utils.tf.Context
 import org.tensorflow.framework.NodeDef
