+++ spark/dl/src/test/scala/com/intel/analytics/bigdl/nn/tf/DecodeImageSpec.scala
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.intel.analytics.bigdl.nn.ops
+package com.intel.analytics.bigdl.nn.tf
 
 import com.intel.analytics.bigdl.utils.tf.TFRecordIterator
 import org.scalatest.{FlatSpec, Matchers}
